pyproject.toml
[project]
name = "polling_agent"
version = "0.1.0"
description = "Polling_agent using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.121.1,<1.0.0",
    "pydantic",
    "streamlit",
    "python-dotenv",
    "tweepy",
    "requests",
]

[project.scripts]
polling_agent = "polling_agent.main:run"
run_crew = "polling_agent.main:run"
train = "polling_agent.main:train"
replay = "polling_agent.main:replay"
test = "polling_agent.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
